Move firebase function to FirebaseManager and login functions to LoginManager
